backend/requirements/requirements.txt
Django==3.2.19
pytz==2023.3
sqlparse==0.4.4
requests==2.26.0
djangorestframework==3.12.4
pillow==9.5.0
djoser==2.1.0
django-filter==23.2
drf-extra-fields==3.5.0
filetype==1.2.0
python-dotenv==0.21.1
gunicorn==20.0.4